Move `DatabaseAdapter` Impl
