+++ src/config/store.rs
@@ -1,4 +1,8 @@
+use anyhow::Result;
 use serde::{Deserialize, Serialize};
+use tracing::warn;
+
+use crate::{config::Config, store::adapters::postgres};
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash, Deserialize, Serialize)]
 #[serde(rename_all = "lowercase")]
@@ -8,4 +12,16 @@ pub enum DatabaseAdapter {
 
     /// PostgreSQL database adapter
     Postgres,
+}
+
+impl DatabaseAdapter {
+    pub async fn migrate(&self, config: &Config) -> Result<()> {
+        match self {
+            Self::Postgres => postgres::migrate(config).await,
+            Self::Sqlite => {
+                warn!("Standalone migration not supported for SQLite");
+                Ok(())
+            }
+        }
+    }
 }
